requirements.txt
langchain
python-dotenv
langchain_community
pypdf
arxiv
pymupdf
wikipedia
langchain_text_splitters
langchain_openai
langchain_ollama
chromadb
langchain_huggingface
sentence_transformers
langchain_chroma
streamlit
langchain_groq
sse_starlette
duckduckgo-search
youtube-transcript-api
validators
unstructured
pytube
numexpr
cassio
tiktoken
PyPDF2